Project5_TexttoSQL_Query/requirements.txt
streamlit==1.32.2
google-generativeai==0.4.1
python-dotenv